Refactor: 메인 여행 탭 메뉴 폴더 구조 변경
