+++ src/components/main/planList/PlanCardList.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function PlanCardList() {
+  return (
+    <div>
+      <div></div>
+    </div>
+  );
+}
